+++ packages/pangea-sdk/pangea/services/file_scan.py
@@ -72,7 +72,6 @@ class FileScan(ServiceBase):
     """
 
     service_name = "file-scan"
-    version = "v1"
 
     def file_scan(
         self,